pyproject.toml
[tool.poetry]
name = "awsmodel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
botocore = "^1.35.69"
pydantic = "^2.10.1"
pydantic-settings = "^2.6.1"
PyYAML = "^6.0.2"
motor = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
